move custom actions used in tests to conftest.py
